refactoring avatar helper
